# docs/api/openapi.v2.json
{
  "components": {
    "schemas": {
...
      "ExactShowSlugResponseV2": {
        "additionalProperties": false,
        "properties": {
          "show": {
            "$ref": "#/components/schemas/ExactShowSlugV2"
          }
        },
        "required": [
          "show"
        ],
        "title": "ExactShowSlugResponseV2",
        "type": "object"
      },
      "ExactShowSlugV2": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "slug": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "slug"
        ],
        "title": "ExactShowSlugV2",
        "type": "object"
      },
...
      "ShowSlugProblemDetailV2": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "retry_after_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retry After Ms"
          },
          "retryable": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retryable"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "trace_id": {
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "status",
          "message",
          "trace_id",
          "request_id"
        ],
        "title": "ShowSlugProblemDetailV2",
        "type": "object"
      },
      "ShowSlugProblemResponseV2": {
        "additionalProperties": false,
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/ShowSlugProblemDetailV2"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ShowSlugProblemResponseV2",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "InternalAdminBearer": {
        "bearerFormat": "JWT",
        "description": "An allowlisted Supabase admin JWT or a TRR-signed internal-admin JWT. The backend applies the same InternalAdminUser authorization policy to both.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Backend API for The Reality Report - reality TV data and surveys",
    "title": "The Reality Report API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v2/admin/shows/exact-slug/{slug}": {
      "get": {
        "operationId": "getAdminShowByExactSlugV2",
        "parameters": [
          {
            "description": "A show slug. Matching is exact after lowercase normalization.",
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "maxLength": 160,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9]+(?:-+[A-Za-z0-9]+)*$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExactShowSlugResponseV2"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShowSlugProblemResponseV2"
                }
              }
            },
            "description": "The show slug is invalid."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShowSlugProblemResponseV2"
                }
              }
            },
            "description": "The show was not found."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShowSlugProblemResponseV2"
                }
              }
            },
            "description": "The exact show-slug read failed."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShowSlugProblemResponseV2"
                }
              }
            },
            "description": "The show store is unavailable."
          }
        },
        "security": [
          {
            "InternalAdminBearer": []
          }
        ],
        "summary": "Get a show by its exact stored slug",
        "tags": [
          "admin-show-slugs-v2"
        ]
      }
    },
...
  }
}